pyproject.toml
[tool.poetry]
name = "cloud-sync"
version = "0.1.0"
description = "Cloud based account provisioning with Azure Automate"
authors = ["Erik Doytchinov <erik.doytchinov@zivver.com>"]
readme = "README.md"

[tool.poetry.scripts]
cloud-sync = "cloud_sync.main:main"

[tool.poetry.dependencies]
python = "^3.8.0"
pydantic = "^2.7.1"
tqdm = "^4.66.4"
httpx = "^0.27.0"
eval-type-backport = "^0.2.0"
tenacity = "^8.3.0"
colorama = "^0.4.6"
coloredlogs = "^15.0.1"
pytest-bdd = "^7.2.0"
pytest = "^8.2.2"
pydantic-settings = "^2.3.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
python-dotenv = "^1.0.1"
pytest = "^8.2.0"
poethepoet = "^0.26.1"
ruff-lsp = "^0.0.53"
curlify2 = "^2.0.0"
pytest-asyncio = "^0.23.7"
pytest-mock = "^3.14.0"

[tool.poe]
include = "poe.toml"

[tool.ruff.lint.per-file-ignores]
# Unused import in pytest files is okay because fixtures
"tests/**/*.py" = ["F401", "F811"]

[pytest]
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
